requirements.txt
python-dotenv==1.0.1
Django==5.0.6
djangorestframework==3.15.
django-filter==24.2
markdown==3.6
psycopg==3.2.1
psycopg-binary==3.2.1
ipython==8.26.0
django-extensions==3.2.3
celery==5.4.0
flower==2.0.1
gevent==24.2.1
redis==5.0.7
django-celery-beat==2.6.0 
django-celery-results==2.5.1
channels==4.1.0
